use sunburst mapping function
