+++ components/pages/discovery/Charts.tsx
@@ -30,156 +30,9 @@ import { toArrangerV3Filter } from 'global/utils/arrangerFilter';
 import { useMemo } from 'react';
 import useFiltersContext from '../file-repository/hooks/useFiltersContext';
 import { addInFilters } from '../file-repository/utils';
+import { mapFromCodeToCancerType } from './cancerTypeMapping';
 import Card from './components/Card';
 import { commonStyles } from './components/common';
-
-/**
- * map
- * inner => outer
- * clicking inner ring filters with outer values
- *
- * our dynamic data from api is the property key
- */
-
-export const cancerTypeCodeMapping = {
-  C00: 'Lip, Oral Cavity & Pharynx',
-  C01: 'Lip, Oral Cavity & Pharynx',
-  C02: 'Lip, Oral Cavity & Pharynx',
-  C03: 'Lip, Oral Cavity & Pharynx',
-  C04: 'Lip, Oral Cavity & Pharynx',
-  C05: 'Lip, Oral Cavity & Pharynx',
-  C06: 'Lip, Oral Cavity & Pharynx',
-  C07: 'Lip, Oral Cavity & Pharynx',
-  C08: 'Lip, Oral Cavity & Pharynx',
-  C09: 'Lip, Oral Cavity & Pharynx',
-  C10: 'Lip, Oral Cavity & Pharynx',
-  C11: 'Lip, Oral Cavity & Pharynx',
-  C12: 'Lip, Oral Cavity & Pharynx',
-  C13: 'Lip, Oral Cavity & Pharynx',
-  C14: 'Lip, Oral Cavity & Pharynx',
-  C15: 'Esophagus',
-  C16: 'Stomach',
-  C17: 'Small Intestine',
-  C18: 'Colorectal',
-  C19: 'Colorectal',
-  C20: 'Colorectal',
-  C21: 'Colorectal',
-  C22: 'Liver & Intrahepatic Bile Duct',
-  C23: 'Gallbladder & Other Biliary',
-  C24: 'Gallbladder & Other Biliary',
-  C25: 'Pancreas',
-  C26: 'Other Digestive Organs',
-  C30: 'Nasal Cavity & Sinuses',
-  C31: 'Nasal Cavity & Sinuses',
-  C32: 'Larynx',
-  C33: 'Lung & Bronchus',
-  C34: 'Lung & Bronchus',
-  C37: 'Thymus, Heart, Mediastinum & Pleura',
-  C38: 'Thymus, Heart, Mediastinum & Pleura',
-  C39: 'Other Respiratory Organs',
-  C40: 'Bone & Articular Cartilage',
-  C41: 'Bone & Articular Cartilage',
-  C43: 'Melanoma of Skin',
-  C44: 'Other Skin',
-  C45: 'Mesothelioma',
-  C46: "Kaposi's Sarcoma",
-  C47: 'Soft Tissue Sarcomas',
-  C49: 'Soft Tissue Sarcomas',
-  C50: 'Breast',
-  C51: 'Vulva & Vagina',
-  C52: 'Vulva & Vagina',
-  C53: 'Cervix Uteri',
-  C54: 'Corpus Uteri & Uterus NOS',
-  C55: 'Corpus Uteri & Uterus NOS',
-  C56: 'Ovary',
-  C57: 'Other Female Genital Organs',
-  C58: 'Placenta',
-  C60: 'Penis',
-  C61: 'Prostate',
-  C62: 'Testis',
-  C63: 'Other Male Genital Organs',
-  C64: 'Kidney & Other Urinary Organs',
-  C65: 'Kidney & Other Urinary Organs',
-  C66: 'Kidney & Other Urinary Organs',
-  C67: 'Bladder',
-  C68: 'Other Urinary Organs',
-  C69: 'Eye & Adnexa',
-  C70: 'Brain & Other CNS',
-  C71: 'Brain & Other CNS',
-  C72: 'Brain & Other CNS',
-  C73: 'Thyroid',
-  C74: 'Other Endocrine Glands',
-  C75: 'Other Endocrine Glands',
-  C76: 'Other Ill-Defined Sites',
-  C77: 'Lymph Nodes',
-  C78: 'Secondary Malignant Neoplasms',
-  C79: 'Secondary Malignant Neoplasms',
-  C80: 'Malignant Neoplasm, Unknown Primary',
-  C81: 'Hodgkin Lymphoma',
-  C82: 'Non-Hodgkin Lymphoma',
-  C83: 'Non-Hodgkin Lymphoma',
-  C84: 'Non-Hodgkin Lymphoma',
-  C85: 'Non-Hodgkin Lymphoma',
-  C86: 'Non-Hodgkin Lymphoma',
-  C88: 'Malignant Immunoproliferative Diseases',
-  C90: 'Multiple Myeloma & Plasmacytoma',
-  C91: 'Leukemias',
-  C92: 'Leukemias',
-  C93: 'Leukemias',
-  C94: 'Leukemias',
-  C95: 'Leukemias',
-  C96: 'Other & Unspecified Lymphoid, Hematopoietic & Related Tissue',
-  C97: 'Multiple Independent Primary Sites',
-  D00: 'Carcinoma in Situ',
-  D01: 'Carcinoma in Situ',
-  D02: 'Carcinoma in Situ',
-  D03: 'Carcinoma in Situ',
-  D04: 'Carcinoma in Situ',
-  D05: 'Carcinoma in Situ',
-  D06: 'Carcinoma in Situ',
-  D07: 'Carcinoma in Situ',
-  D08: 'Carcinoma in Situ',
-  D09: 'Carcinoma in Situ',
-  D10: 'Benign Neoplasms',
-  D11: 'Benign Neoplasms',
-  D12: 'Benign Neoplasms',
-  D13: 'Benign Neoplasms',
-  D14: 'Benign Neoplasms',
-  D15: 'Benign Neoplasms',
-  D16: 'Benign Neoplasms',
-  D17: 'Benign Neoplasms',
-  D18: 'Benign Neoplasms',
-  D19: 'Benign Neoplasms',
-  D20: 'Benign Neoplasms',
-  D21: 'Benign Neoplasms',
-  D22: 'Benign Neoplasms',
-  D23: 'Benign Neoplasms',
-  D24: 'Benign Neoplasms',
-  D25: 'Benign Neoplasms',
-  D26: 'Benign Neoplasms',
-  D27: 'Benign Neoplasms',
-  D28: 'Benign Neoplasms',
-  D29: 'Benign Neoplasms',
-  D30: 'Benign Neoplasms',
-  D31: 'Benign Neoplasms',
-  D32: 'Benign Neoplasms',
-  D33: 'Benign Neoplasms',
-  D34: 'Benign Neoplasms',
-  D35: 'Benign Neoplasms',
-  D36: 'Benign Neoplasms',
-  D37: 'Neoplasms of Uncertain or Unknown Behaviour',
-  D38: 'Neoplasms of Uncertain or Unknown Behaviour',
-  D39: 'Neoplasms of Uncertain or Unknown Behaviour',
-  D40: 'Neoplasms of Uncertain or Unknown Behaviour',
-  D41: 'Neoplasms of Uncertain or Unknown Behaviour',
-  D42: 'Neoplasms of Uncertain or Unknown Behaviour',
-  D43: 'Neoplasms of Uncertain or Unknown Behaviour',
-  D44: 'Neoplasms of Uncertain or Unknown Behaviour',
-  D45: 'Neoplasms of Uncertain or Unknown Behaviour',
-  D46: 'Neoplasms of Uncertain or Unknown Behaviour',
-  D47: 'Neoplasms of Uncertain or Unknown Behaviour',
-  D48: 'Neoplasms of Uncertain or Unknown Behaviour',
-};
 
 const getAgeAtDiagnosisFilter = (key, field) => {
   // ranges from query are less than 18, 18 => 65, 65+
@@ -407,7 +260,7 @@ const ChartsLayout = () => {
         >
           <SunburstChart
             fieldName="primary_diagnosis__cancer_type_code"
-            mapping={cancerTypeCodeMapping}
+            mapping={mapFromCodeToCancerType}
             handlers={{
               onClick: (config) => {
                 const setFilter = chartFilter('primary_diagnosis.cancer_type_code');
